requirements.txt
newsapi-python
openai
ibm-watson
ibm-cloud-sdk-core
streamlit
dotenv